move apollo client into plugins
