+++ src/taskpane/Layout/Layout.tsx
@@ -14,7 +14,7 @@ import {
 } from '@fluentui/react';
 import React, { ReactElement } from 'react';
 import { useLogoutMutation } from '../../generated/graphql';
-import client from '../../utils/apolloClient';
+import client from '../../plugins/apollo/apolloClient';
 import CitationStyle from '../pages/citationStyle';
 import Dashboard from '../pages/dashboard';
 import { useCiteSupport } from '../contexts/CiteSupportContext';
